ui: redesign settings * split categories to screens * handle `locale` `configChanges` * replace some deprecated things
