+++ app/src/main/AndroidManifest.xml
@@ -14,11 +14,12 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/Theme.LibreSudoku"
-        tools:targetApi="tiramisu">
+        tools:targetApi="tiramisu"
+        >
         <activity
             android:name=".MainActivity"
             android:exported="true"
-            android:label="@string/app_name"
+            android:configChanges="locale|layoutDirection"
             android:screenOrientation="portrait"
             android:theme="@style/Theme.LibreSudoku">
             <intent-filter>
